benchmarks/requirements.txt
pyyaml
tiktoken
anthropic>=0.40.0
plotly
kaleido
